requirements.txt
PySide6>=6.7.2,<6.9
pytesseract>=0.3.13
Pillow>=10.4
qhotkey>=0.3.0; platform_system == "Windows"
